pyproject.toml
[tool.poetry]
name = "mongopycrud"
version = "0.1.0"
description = ""
authors = ["luis <vnunes30@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pymongo = "^4.3.3"
fastapi = {extras = ["all"], version = "^0.95.0"}
python-dotenv = "^1.0.0"
pydantic = "^1.10.7"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
taskipy = "^1.10.4"
isort = "^5.12.0"
black = "^23.3.0"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
line_length = 79

[tool.taskipy.tasks]
lint = "black --check --diff . && isort --check --diff ."
pre_test = "task lint"
test = "pytest -s -x --cov=mongoPyCrud -vv"
post_test = "coverage html"
run = " uvicorn main:app --reload"